+++ src/Http/Controllers/Controller.php
@@ -2,12 +2,10 @@
 
 namespace WeiJuKeJi\LaravelDictionary\Http\Controllers;
 
-use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
-use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Routing\Controller as BaseController;
-use WeiJuKeJi\LaravelDictionary\Support\ApiResponse;
+use WeiJuKeJi\LaravelDictionary\Http\Controllers\Concerns\RespondsWithApi;
 
-class Controller extends BaseController
+abstract class Controller extends BaseController
 {
-    use AuthorizesRequests, ValidatesRequests, ApiResponse;
+    use RespondsWithApi;
 }