+++ src/Http/Requests/DictionaryItem/DictionaryItemStoreRequest.php
@@ -17,13 +17,15 @@ class DictionaryItemStoreRequest extends FormRequest
         return [
             'parent_key' => ['required', 'string', 'max:100'],
             'item_key' => [
-                'required',
+                'required_if:auto_generate_key,false',
+                'nullable',
                 'string',
                 'max:100',
                 Rule::unique('dictionary_items', 'item_key')
                     ->where('parent_key', $this->input('parent_key'))
             ],
             'item_value' => ['required', 'string', 'max:200'],
+            'auto_generate_key' => ['sometimes', 'boolean'],
             'sort_order' => ['sometimes', 'integer', 'min:0'],
             'is_enabled' => ['sometimes', 'boolean'],
         ];
@@ -35,6 +37,7 @@ class DictionaryItemStoreRequest extends FormRequest
             'parent_key' => '父分类Key',
             'item_key' => '字典项Key',
             'item_value' => '字典项值',
+            'auto_generate_key' => '自动生成字典Key',
             'sort_order' => '排序',
             'is_enabled' => '是否启用',
         ];
@@ -51,6 +54,10 @@ class DictionaryItemStoreRequest extends FormRequest
     {
         if ($this->has('is_enabled')) {
             $this->merge(['is_enabled' => $this->boolean('is_enabled')]);
+        }
+
+        if ($this->has('auto_generate_key')) {
+            $this->merge(['auto_generate_key' => $this->boolean('auto_generate_key')]);
         }
     }
 }